requirements.txt
 djangorestframework==3.15.2 
 mysqlclient==2.2.4 
 python-dotenv==1.0.1 
 drf-yasg==1.21.7